+++ CHANGELOG.md
@@ -30,6 +30,8 @@ You can find our backwards-compatibility policy [here](https://github.com/hynek/
 - `structlog.processors.LogfmtRenderer` now escapes backslashes and double quotes.
   [#594](https://github.com/hynek/structlog/pull/594)
 
+- `structlog.testing.CapturingLogger` now handles `bind`, `unbind` and `new` as expected by `BoundLogger` returning itself.
+  [#604](https://github.com/hynek/structlog/pull/604)
 
 
 ## [24.1.0](https://github.com/hynek/structlog/compare/23.3.0...24.1.0) - 2024-01-08